Dummy visualization PR

+++ src/main/java/hu/oe/nik/szfmv17t/environment/utils/XmlParser.java
@@ -136,8 +136,8 @@ public class XmlParser {
         //where a * b is the scalarProduct
         
         //Our zero degree will be the horizontal right:
-        int defaultX = 1;
-        int defaultY = 0;
+        double defaultX = 1;
+        double defaultY = 0;
 
         double transformedX = m11*defaultX + m12*defaultY;
         double transformedY = m21*defaultX + m22*defaultY;
@@ -146,7 +146,7 @@ public class XmlParser {
         
         double defaultVectorLength = Math.sqrt(defaultX * defaultX + defaultY * defaultY);
         double transformedVectorLength = Math.sqrt(transformedX * transformedX + transformedY * transformedY);
-        
+   
         double angleInRad = Math.acos(scalarProduct / (defaultVectorLength * transformedVectorLength));
         if (transformedY < 0) {
             angleInRad = 2*Math.PI - angleInRad;
@@ -154,6 +154,7 @@ public class XmlParser {
         //If angle is NaN as a result of transformedVectorLength=0, Math.round() returns 0. It is correct in our cases.
         angleInRad = Math.round(angleInRad * 100.0) / 100.0;     
         return angleInRad;
+
     }
 
     public List<IWorldObject> getWorldObjects() {
